rocketnet-sites/requirements.txt
fastmcp>=2.12.0
httpx>=0.24.0
python-dotenv>=1.0.0

# Install shared package (for local development)
# pip install -e ../rocketnet-shared